+++ R/time_seq_id.R
@@ -102,7 +102,8 @@ time_seq_id <- function(x, time_by = NULL, threshold = 1,
     dt <- data.table::data.table(x = telapsed, group_id = GRP_group_id(g))
     group_id_col <- names(dt)[names(dt) == "group_id"]
     over_threshold <- dt[, ("over") :=
-                           roll_time_threshold(get("x"), threshold = threshold,
+                           roll_time_threshold(get("x"),
+                                               threshold = threshold,
                                                switch_on_boundary = switch_on_boundary),
                          by = group_id_col][["over"]]
   }