+++ R/timeplyr-package.R
@@ -1,5 +1,4 @@
 ## usethis namespace: start
 #' @useDynLib timeplyr, .registration = TRUE
-#' @importFrom Rcpp sourceCpp
 ## usethis namespace: end
 NULL