+++ tests/testthat/test-is_whole_number.R
@@ -41,8 +41,7 @@ testthat::test_that("Whole numbers", {
   testthat::expect_false(is_whole_number(c(-Inf, y)))
 
   testthat::expect_true(
-    is_whole_number(c(110 * 10^200, 1.1 * 100 * 10^200, sqrt(.Machine$double.eps)/10),10^9 + 0.02,
-                    10^9 + 2)
+    is_whole_number(c(110 * 10^200, 1.1 * 100 * 10^200, 10^9 + 2))
   )
   testthat::expect_false(is_whole_number(10^9 + 0.02))